% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{append_details}
\alias{append_details}
\alias{set_details}
\title{Internal helper function: Append and set elements of a details list}
\usage{
append_details(details, index, further)

set_details(details, index, value)
}
\arguments{
\item{details}{A list of details to be requested from
\code{baskwrap::get_details.fujikawa_x} (will be supplied to the function as
the \code{which_details} argument).}

\item{index}{Indicates which element of \code{details} should be appended.}

\item{further}{A list of further parameters to be appended.}

\item{value}{A character string, \code{details[[index]]} will be assigned \code{value}.}
}
\value{
The updated list of \code{details}.
}
\description{
\code{append_details} takes the element \code{details[["index"]]} and appends it with
the list \code{further}. \code{set_details} takes the element \code{details[["index"]]} and
sets it to \code{value}.
}
