% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{append_which_detail_params}
\alias{append_which_detail_params}
\title{Internal helper function: Append which_details list in the detail_params list}
\usage{
append_which_detail_params(detail_params, further)
}
\arguments{
\item{detail_params}{A named list of parameters that need to be supplied to
\code{get_details()}. If \code{p1} exists, \code{detail_params$p1} is replaced by \code{p1} for
calculating  power resp. ECD and by \code{p2} for calculating FWER.}

\item{further}{A list of further params.}
}
\value{
The updated list of \code{detail_params}.
}
\description{
Takes the argument \code{detail_params$which_details} and appends it with the
list \code{further}.
}
