% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basket.R
\docType{class}
\name{basket}
\alias{basket}
\title{Class for a generic basket trial design.}
\description{
This is the generic class defining a basket trial design. A basket trial
design is study design for single-arm studies with binary endpoints which are
subdivided into \code{k} disease subtypes, the so-called baskets, with \code{n}
individuals per basket.
}
\section{Slots}{

\describe{
\item{\code{k}}{A positive integer denoting the number of baskets.}

\item{\code{n}}{A positive integer denoting the number of individuals per basket.}

\item{\code{tuning_params}}{A list of real numbers denoting parameters for tuning the
design.}
}}

