% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{decide}
\alias{decide}
\title{Generic function for deciding which baskets are considered active
("go" decision) and which baskets are considered inactive. The decision is
based on the number of responses in the baskets.}
\usage{
decide(x)
}
\arguments{
\item{x}{numeric vector of number of responses in each basket}
}
\value{
a logical vector of decisions for each basket
}
\description{
Generic function for deciding which baskets are considered active
("go" decision) and which baskets are considered inactive. The decision is
based on the number of responses in the baskets.
}
