% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fujikawa.R
\docType{class}
\name{fujikawa}
\alias{fujikawa}
\title{Class for Fujikawa's basket trial design.}
\description{
This is the generic class defining the basket trial design suggested by
Fujikawa et al., 2020.
}
\section{Slots}{

\describe{
\item{\code{tuning_params}}{A list of two real numbers of the form
\code{list(epsilon = , tau = )} denoting the tuning parameters.}

\item{\code{prior_params}}{A list of two real numbers of the form \code{list(a = , b = )}
denoting shape parameters for the beta distribution $\mathrm{Beta}(a, b)$ which
models the prior distribution of the response probability. So far, only
independent identically distributed baskets are implemented. Real number
vectors for \code{a} and \code{b} corresponding to individual prior distributions for
different baskets may be implemented in the future.}

\item{\code{weights_fun}}{A string denoting the name of the weights function. Default
is \code{"jensen_shannon"}, meaning that the similarity of baskets is based on
the Jensen-Shannon divergence.}
}}

