% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_details_for_two_scenarios}
\alias{get_details_for_two_scenarios}
\title{Internal helper function: Get details for two response scenarios}
\usage{
get_details_for_two_scenarios(design, x, detail_params, p1, p2)
}
\arguments{
\item{design}{An object of class \code{fujikawa} created by the function
\code{baskwrap::setup_fujikawa_x} or \code{basksim::setup_fujikawa}.}

\item{x}{A named list, the design's tuning parameters to be optimized.}

\item{detail_params}{A named list of parameters that need to be supplied to
\code{get_details()}. If \code{p1} exists, \code{detail_params$p1} is replaced by \code{p1} for
calculating  power resp. ECD and by \code{p2} for calculating FWER.}

\item{p1}{A numeric, response scenario for calculating power resp. ECD. This
can also be left \code{NULL}, then the function goes looking in
\code{detail_params$p1}.}

\item{p2}{A numeric, response scenario for calculating FWER, default is the
global null scenario.}
}
\value{
A list of two lists containing return values of \code{get_details} calls.
}
\description{
Internal helper function: Get details for two response scenarios
}
