% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_p1s}
\alias{get_p1s}
\title{Generate an array of possible outcome scenarios}
\usage{
get_p1s(k, p0, p1, by = 1)
}
\arguments{
\item{k}{The number of baskets.}

\item{p0}{A common probability under the null hypothesis.}

\item{p1}{A numeric, the true response probability for the active baskets}

\item{by}{A numeric, increment of the scenario sequence, default: 1.}
}
\value{
A numeric array.
}
\description{
This functions generates an array of possible outcome scenarios. Each row
of the array is a scenario vector. The array starts with \code{k} inactive
baskets and 0 active baskets and then increments in steps of \code{by} up to
\code{k} active baskets (or the next smaller multiple of \code{by}).
}
\examples{
p1s <- get_p1s(k = 3, p0 = 0.2, p1 = 0.5)
}
