% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_scenarios}
\alias{get_scenarios}
\title{Generate a scenario of possible outcomes}
\usage{
get_scenarios(k, p0, p1, by = 1)
}
\arguments{
\item{k}{The number of baskets.}

\item{p0}{A common probability under the null hypothesis.}

\item{p1}{A numeric, the true response probability for the active baskets}

\item{by}{A numeric, increment of the scenario sequence, default: 1.}
}
\value{
A list with components \code{k}, \code{p0} and \code{p1s},
where \code{p1s} is generated by the the function \code{get_p1s}.
}
\description{
A wrapper of \code{get_p1s()} that also returns the input values \code{k} and
\code{p0} as part of a list.
}
\examples{
scenario <- get_scenarios(k = 3, p0 = 0.2, p1 = 0.5)
}
