% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_trace_info}
\alias{get_trace_info}
\title{Internal helper function: Decide whether to record a trace and what path to
use}
\usage{
get_trace_info(trace, type = "trace")
}
\arguments{
\item{trace}{A parameter that can be \code{FALSE}, \code{TRUE} or a file path.}

\item{type}{A character that designates the name of the input.}
}
\value{
A list with components \code{rec} and \code{path}.
}
\description{
Internal helper function: Decide whether to record a trace and what path to
use
}
