% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{io_val_p1}
\alias{io_val_p1}
\title{Internal helper function: Input validation for p1}
\usage{
io_val_p1(detail_params, p1)
}
\arguments{
\item{detail_params}{A named list of parameters that need to be supplied to
\code{get_details()}. If \code{p1} exists, \code{detail_params$p1} is replaced by \code{p1} for
calculating  power resp. ECD and by \code{p2} for calculating FWER.}

\item{p1}{A numeric, response scenario for calculating power resp. ECD. This
can also be left \code{NULL}, then the function goes looking in
\code{detail_params$p1}.}
}
\value{
The updated list of \code{detail_params}.
}
\description{
Returns \code{detail_params} with checked element \code{detail_params$p1 <- p1},
depending on whether \code{p1} is not \code{NULL} or \code{detail_params$p1} is not \code{NULL}.
Returns an error message if both are NULL.
}
