% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{params_main}
\alias{params_main}
\title{List of parameters used across the package}
\arguments{
\item{design}{An object of class \code{fujikawa} created by the function
\code{baskwrap::setup_fujikawa_x} or \code{basksim::setup_fujikawa}.}

\item{x}{A named list, the design's tuning parameters to be optimized.}

\item{detail_params}{A named list of parameters that need to be supplied to
\code{get_details()}.}
}
\description{
List of parameters used across the package
}
