% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{params_utility_caller}
\alias{params_utility_caller}
\title{List of parameters used for functions calling utility functions}
\arguments{
\item{utility}{A function returning the utility of a parameter combination
\code{x} of the form \code{utility(design, x, detail_params, ...)}, where \code{...} may
further parameters to be supplied to the utility.}

\item{utility_params}{A named list of further parameters that need to be
supplied to  the utility function.}
}
\description{
List of parameters used for functions calling utility functions
}
