% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{u_2ewp}
\alias{u_2ewp}
\title{Utility function: Two-level family-wise power-error function}
\usage{
u_2ewp(design, x, detail_params, xi1, xi2, thresh)
}
\arguments{
\item{design}{An object of class \code{Basket} created by the function
\code{baskexact::setupOneStageBasket}.}

\item{x}{A named list of the design parameters to be optimized.}

\item{detail_params}{A named list of parameters that need to be supplied to
\code{baskexact::toer()} and \code{baskexact::pow()}.}

\item{xi1}{A numeric, \code{xi1} is the penalty for low FWER,}

\item{xi2}{A numeric, \code{xi1 + xi2} is the penalty for high FWER,}

\item{thresh}{A numeric, for high FWER above this treshold we impose harder penalty.}
}
\value{
a numerical, the parameter combination's utility
}
\description{
This utility function is defined as
\deqn{u_{\text{2ewp}}(\boldsymbol\phi,\mathbf{p}) = \mathrm{ewp}(\boldsymbol
\phi,\mathbf{p}) - \left(\xi_1\mathrm{fwer}(\boldsymbol\phi,\mathbf{p}) +
\xi_2(\mathrm{fwer}(\boldsymbol\phi,\mathbf{p}) - \eta)\mathbf 1(
\mathrm{fwer}(\boldsymbol\phi,\mathbf{p}) - \eta) \right)}
where \eqn{\eta\in[0,1]} is a threshold for imposing harder FWER penalty.
}
\examples{
design <- setupOneStageBasket(k = 3, shape1 = 1, shape2 = 1, p0 = 0.2)
u_2ewp(design,
              x = list(lambda = 0.99, epsilon = 2, tau = 0.5),
              detail_params = list(p1 = c(0.5, 0.2, 0.2),
                                   n = 20,
                                   weight_fun = weights_fujikawa,
                                   logbase = exp(1)),
              xi1 = 1, xi2 = 2,
              thresh = 0.1)
}
