% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{u_avg}
\alias{u_avg}
\title{Utility function: Scenario-averaged utility function}
\usage{
u_avg(
  design,
  x,
  detail_params,
  utility,
  utility_params,
  p1s,
  weights_u = rep(1/nrow(p1s), nrow(p1s))
)
}
\arguments{
\item{design}{An object of class \code{Basket} created by the function
\code{baskexact::setupOneStageBasket}.}

\item{x}{A named list of the design parameters to be optimized.}

\item{detail_params}{A named list of parameters that need to be supplied to
\code{baskexact::toer()} and \code{baskexact::pow()}. It must not contain \code{p1},
as this is supplied separately.}

\item{utility}{A function returning the utility of a parameter combination
\code{x}, of the form \code{utility(design, x, detail_params, ...)}, where \code{...} may
further parameters to be supplied to the utility.}

\item{utility_params}{A named list of further parameters that need to be
supplied to  the utility function.}

\item{p1s}{A numeric array in which each row defines a scenario of true response rates
under the alternative hypothesis.}

\item{weights_u}{A numeric vector of weights for calculating the weighted average.}
}
\value{
a numerical, the parameter combination's utility
}
\description{
For a utility function \eqn{u(\cdot, \mathbf p)} and a set of true scenarios
\eqn{\{\mathbf p_i,\ldots\}}, calculate the weighted average utility function
\deqn{ \bar u(x) = \sum_i w_i u(x, \mathbf p_i)}
for a set of weights with \eqn{\sum_i w_i=1}. By default, \eqn{w_i=\frac{1}{|\{\mathbf p_i,\ldots\}|}} for all \eqn{i}.
}
\examples{
design <- setupOneStageBasket(k = 3, shape1 = 1, shape2 = 1, p0 = 0.2)
x <- list(lambda = 0.99, epsilon = 2, tau = 0.5)
detail_params <- list(n = 20,
                      weight_fun = weights_fujikawa,
                      logbase = exp(1))
p1s <- rbind(c(0.2,0.2,0.2), c(0.2,0.2,0.5), c(0.2,0.5,0.5), c(0.5,0.5,0.5))
u_avg(design,
      x = x,
      detail_params = detail_params,
      utility = u_ewp_discont,
      utility_params = list(thresh = 0.05),
      p1s = p1s
      )
u_avg(design,
      x = x,
      detail_params = detail_params,
      utility = u_2ewp,
      utility_params = list(xi1 = 1, xi2 = 2, thresh = 0.1),
      p1s = p1s
      )
}
