% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{u_ewp_discont_bound}
\alias{u_ewp_discont_bound}
\title{Utility function: Discontinuous power function with type-I error penalty and
boundaries on the parameters}
\usage{
u_ewp_discont_bound(design, x, detail_params, alpha, lower, upper)
}
\arguments{
\item{design}{An object of class \code{Basket} created by the function
\code{baskexact::setupOneStageBasket}.}

\item{x}{A named list of the design parameters to be optimized.}

\item{detail_params}{A named list of parameters that need to be supplied to
\code{baskexact::toer()} and \code{baskexact::pow()}.}

\item{lower}{numerical, a vector of lower bounds of the parameters.}

\item{upper}{numerical, a vector of upper bounds of the parameters.}
}
\description{
Utility function: Discontinuous power function with type-I error penalty and
boundaries on the parameters
}
\examples{
design <- setupOneStageBasket(k = 3, shape1 = 1, shape2 = 1, p0 = 0.2)
u_ewp_discont_bound(design,
              x = list(lambda = 0.99, epsilon = 2, tau = 0.5),
              detail_params = list(p1 = c(0.5, 0.2, 0.2),
                                   n = 20,
                                   weight_fun = weights_fujikawa,
                                   logbase = exp(1)),
              thresh = 0.05,
                  lower = list(lambda = 0, epsilon = 1, tau = 0),
                  upper = list(lambda = 1, epsilon = 10, tau = 1))
}
