% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{u_ewp_discont_sim}
\alias{u_ewp_discont_sim}
\title{Utility function: Discontinuous power function with type-I error penalty (simulated)}
\usage{
u_ewp_discont_sim(design, x, detail_params, thresh)
}
\arguments{
\item{design}{An object of class \code{fujikawa} created by the function
\code{basksim::setup_fujikawa}.}

\item{x}{A named list of the design parameters to be optimized.}

\item{detail_params}{A named list of parameters that need to be supplied to
\code{basksim::get_details()}.}
}
\value{
a numerical, the parameter combination's utility
}
\description{
Utility function: Discontinuous power function with type-I error penalty (simulated)
}
\examples{
design <- basksim::setup_fujikawa(k = 3, shape1 = 1, shape2 = 1, p0 = 0.2)
u_ewp_discont_sim(design,
              x = list(lambda = 0.99, epsilon = 2, tau = 0.5),
              detail_params = list(p1 = c(0.5, 0.2, 0.2),
                                   n = 20,
                                   iter = 1000,
                                   logbase = exp(1)),
              thresh = 0.05)
}
