% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{append_details}
\alias{append_details}
\alias{set_details}
\title{Internal helper function: Append and set elements of a details list}
\usage{
append_details(details, index, further)

set_details(details, index, value)
}
\arguments{
\item{further}{A list of further parameters to be appended.}

\item{value}{Any value that can be a list element.}
}
\value{
The updated list of \code{details}.
}
\description{
\code{append_details} takes the element \code{details[["index"]]} and appends it with
the list \code{further}. \code{set_details} takes the element \code{details[["index"]]} and
sets it to \code{value}.
}
