% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_borrowing.R
\name{epsilon_extreme}
\alias{epsilon_extreme}
\title{Extreme borrowing cutoff for epsilon dependent on tau}
\usage{
epsilon_extreme(design, tau, detail_params)
}
\arguments{
\item{design}{An object of class \code{Basket} created by the function
\code{baskexact::setupOneStageBasket}.}

\item{tau}{The optimization parameter \eqn{\tau} of Fujikawa's basket trial
design.}

\item{detail_params}{A named list of parameters containing per-stratum
sizes \code{n} and the base of the logarithm for calculating Jensen-Shannon
divergence \code{logbase}.}
}
\value{
A numeric, the extreme borrowing cutoff for \eqn{\varepsilon}.
}
\description{
Extreme borrowing cutoff for epsilon dependent on tau
}
\examples{
design <- basksim::setup_fujikawa(k = 3,
                                 shape1 = 1,
                                 shape2 = 1,
                                 p0 = 0.2)
detail_params <- list(p1 = c(0.5, 0.2, 0.2),
                             n = 20,
                             logbase = exp(1))
epsilon_extreme(design, tau = 0.5, detail_params)
}
