% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{u_2ewp}
\alias{u_2ewp}
\alias{u_2pow}
\title{Utility functions: Two-level power-error combination functions}
\usage{
u_2ewp(
  design,
  x,
  detail_params,
  p1 = NULL,
  penalty1,
  penalty2,
  threshold,
  report_details = FALSE
)

u_2pow(
  design,
  x,
  detail_params,
  p1 = NULL,
  penalty1,
  penalty2,
  threshold,
  report_details = FALSE
)
}
\arguments{
\item{design}{An object of class \code{fujikawa} created by the function
\code{baskwrap::setup_fujikawa_x} or \code{basksim::setup_fujikawa}.}

\item{x}{A named list, the design's tuning parameters to be optimized.}

\item{detail_params}{A named list of parameters that need to be supplied to
\code{get_details()}.}

\item{p1}{A numeric, response scenario for calculating power and error rate.}

\item{penalty1}{A numeric, \code{penalty1} is the penalty for low FWER,}

\item{penalty2}{A numeric, \code{penalty1 + penalty2} is the penalty for high
FWER.}

\item{threshold}{A numeric, for high FWER above this threshold we impose a
harsher penalty.}

\item{report_details}{A logical, if \code{TRUE}, the function returns the output
of the implicit call to \code{baskwrap::get_details()} as \code{attr(, "details")}.}
}
\value{
A numeric, the parameter combination's utility.
}
\description{
These utility functions combine rewarding power and penalizing TOER by
subtracting TOER from power. In addition, unacceptably high FWER above a
certain threshold receives harsher penalty.
}
\details{
The utility function \eqn{u_{\text{2ewp}}} is defined as
\deqn{u_{\text{2ewp}}(\boldsymbol\phi,\mathbf{p}) = \mathrm{ewp}(\boldsymbol
\phi,\mathbf{p}) - \left(\xi_1\mathrm{fwer}(\boldsymbol\phi,\mathbf{p}) +
\xi_2(\mathrm{fwer}(\boldsymbol\phi,\mathbf{p}) - \eta)\mathbf 1(
\mathrm{fwer}(\boldsymbol\phi,\mathbf{p}) - \eta) \right),}
where \eqn{\eta\in[0,1]} is a threshold for imposing harsher FWER penalty.

The utility function \eqn{u_{\text{2pow}}} is defined analogously as
\deqn{u_{\text{2pow}}(\boldsymbol\phi,\mathbf{p}) = \sum_{i\in R}\mathrm{pow}_i(\boldsymbol
\phi,\mathbf{p}) - \sum_{j\in R^c}\left(\xi_1\mathrm{toer}_j(\boldsymbol\phi,\mathbf{p}) +
\xi_2(\mathrm{toer}_j(\boldsymbol\phi,\mathbf{p}) - \eta)\mathbf 1(
\mathrm{toer}_j(\boldsymbol\phi,\mathbf{p}) - \eta) \right),}
where \eqn{R} and \eqn{R^c} are the sets of active and inactive strata, respectively.
}
\examples{
design <- baskwrap::setup_fujikawa_x(k = 3, shape1 = 1, shape2 = 1,
                                     p0 = 0.2, backend = "exact")
u_2ewp(design,
       x = list(lambda = 0.99, epsilon = 2, tau = 0.5),
       detail_params = list(p1 = c(0.5, 0.2, 0.2),
                            n = 20,
                            weight_fun = baskexact::weights_fujikawa,
                            logbase = exp(1)),
       penalty1 = 1, penalty2 = 2,
       threshold = 0.1)
u_2pow(design,
       x = list(lambda = 0.99, epsilon = 2, tau = 0.5),
       detail_params = list(p1 = c(0.5, 0.2, 0.2),
                            n = 20,
                            weight_fun = baskexact::weights_fujikawa,
                            logbase = exp(1)),
       penalty1 = 1, penalty2 = 2,
       threshold = 0.1)
}
