% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{u_bnd}
\alias{u_bnd}
\title{Utility function with boundaries on the parameters}
\usage{
u_bnd(design, x, detail_params, utility, utility_params, lower, upper)
}
\arguments{
\item{design}{An object of class \code{fujikawa} created by the function
\code{baskwrap::setup_fujikawa_x} or \code{basksim::setup_fujikawa}.}

\item{x}{A named list of the design parameters to be optimized.}

\item{detail_params}{A named list of parameters that need to be supplied to
\code{get_details()}.}

\item{lower}{numerical, a vector of lower bounds of the parameters.}

\item{upper}{numerical, a vector of upper bounds of the parameters.}
}
\value{
a numeric, the parameter combination's utility
}
\description{
Utility function with boundaries on the parameters
}
\examples{
design <- baskwrap::setup_fujikawa_x(k = 3, shape1 = 1, shape2 = 1,
                                     p0 = 0.2, backend = "exact")
u_bnd(design,
      x = c(lambda = 0.99, epsilon = 2, tau = 0.5),
      detail_params = list(p1 = c(0.5, 0.2, 0.2),
                           n = 20,
                           weight_fun = baskexact::weights_fujikawa,
                           logbase = exp(1)),
      utility = u_ewp_discont,
      utility_params = list(thresh = 0.05),
      lower = c(lambda = 0, epsilon = 1, tau = 0),
      upper = c(lambda = 1, epsilon = 10, tau = 1))
}
